server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.12

sqlalchemy>=2.0.36
alembic>=1.14.0
python-dotenv==1.0.0

pymupdf>=1.24.10  # PDF to image
Pillow>=10.3.0
python-magic

psycopg2-binary>=2.9.9

google-generativeai>=0.8.0

python-dateutil==2.8.2
pytz==2023.3
